pyproject.toml
[project]
name = "strava-brrr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "stravalib>=2.4",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
